Set up fontend webpack dev and prod config
This sets up a basic configuration for tailwind as well as the
development and production modes. The production mode is more complex
due to the presences of purge css, which looks through the reason and
html files and cuts out any unused css classes in order to trim down the
css file built with tailwind.